+++ test/integration/mongo_libs/crt_coll_inst.py
@@ -43,6 +43,12 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
+        test_coll_auth_mech3 -> Test with authentication mechanism MONGODB-CR.
+        test_coll_auth_mech2 -> Test with authentication mechanism.
+        test_coll_auth_mech -> Test with authentication mechanism.
+        test_repset_auth_mech3 -> Test with auth mechanism - MONGODB-CR.
+        test_repset_auth_mech2 -> Test with authentication mechanism.
+        test_repset_auth_mech -> Test with authentication mechanism.
         test_coll2 -> Test the mongo_class.Coll class.
         test_coll -> Test the mongo_class.Coll class.
         test_repsetcoll2 -> Test the mongo_class.RepSetColl class.
@@ -66,6 +72,105 @@ class UnitTest(unittest.TestCase):
         self.cfg = gen_libs.load_module(self.config_name, self.config_dir)
         self.dbn = "admin"
         self.coll = "system.users"
+
+    def test_coll_auth_mech3(self):
+
+        """Function:  test_coll_auth_mech3
+
+        Description:  Test with authentication mechanism - MONGODB-CR.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = None
+        self.cfg.auth_mech = "MONGODB-CR"
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertEqual((mongo.auth_db, mongo.auth_mech),
+                         (self.cfg.auth_db, self.cfg.auth_mech))
+
+    def test_coll_auth_mech2(self):
+
+        """Function:  test_coll_auth_mech2
+
+        Description:  Test with authentication mechanism.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = None
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertTrue(isinstance(mongo, mongo_class.Coll))
+
+    def test_coll_auth_mech(self):
+
+        """Function:  test_coll_auth_mech
+
+        Description:  Test with authentication mechanism.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = None
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertEqual((mongo.auth_db, mongo.auth_mech),
+                         (self.cfg.auth_db, self.cfg.auth_mech))
+
+    def test_repset_auth_mech3(self):
+
+        """Function:  test_repset_auth_mech3
+
+        Description:  Test with authentication mechanism - MONGODB-CR.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = self.cfg.host + ":" + str(self.cfg.port)
+        self.cfg.auth_mech = "MONGODB-CR"
+
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertEqual((mongo.auth_db, mongo.auth_mech),
+                         (self.cfg.auth_db, self.cfg.auth_mech))
+
+    def test_repset_auth_mech2(self):
+
+        """Function:  test_repset_auth_mech2
+
+        Description:  Test with authentication mechanism.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = self.cfg.host + ":" + str(self.cfg.port)
+
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertTrue(isinstance(mongo, mongo_class.RepSetColl))
+
+    def test_repset_auth_mech(self):
+
+        """Function:  test_repset_auth_mech
+
+        Description:  Test with authentication mechanism.
+
+        Arguments:
+
+        """
+
+        self.cfg.repset_hosts = self.cfg.host + ":" + str(self.cfg.port)
+
+        mongo = mongo_libs.crt_coll_inst(self.cfg, self.dbn, self.coll)
+
+        self.assertEqual((mongo.auth_db, mongo.auth_mech),
+                         (self.cfg.auth_db, self.cfg.auth_mech))
 
     def test_auth_db_coll2(self):