+++ test/unit/mongo_class/db_init.py
@@ -41,6 +41,8 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
+        test_config_attr3 -> Test with SCRAM-SHA-1 setting.
+        test_config_attr2 -> Test with MONGODB-CR setting.
         test_db_attr -> Test db attribute passed.
         test_no_db_name_attr -> Test no db_name attribute passed.
         test_db_name_attr -> Test db_name attribute passed.
@@ -86,6 +88,40 @@ class UnitTest(unittest.TestCase):
         self.conn_list = [self.host + ":" + str(self.port)]
         self.conf_file = "Config file"
         self.db_name = "MyDatabase"
+        self.auth_mech = "MONGODB-CR"
+        self.auth_mech2 = "SCRAM-SHA-1"
+        self.config2 = {key1 + key2: self.japd,
+                        "authMechanism": self.auth_mech2}
+
+    def test_config_attr3(self):
+
+        """Function:  test_config_attr3
+
+        Description:  Test with SCRAM-SHA-1 setting.
+
+        Arguments:
+
+        """
+
+        mongo = mongo_class.DB(self.name, self.user, self.japd,
+                               self.host, self.port, auth_mech=self.auth_mech2)
+
+        self.assertEqual(mongo.config, self.config2)
+
+    def test_config_attr2(self):
+
+        """Function:  test_config_attr2
+
+        Description:  Test with MONGODB-CR setting.
+
+        Arguments:
+
+        """
+
+        mongo = mongo_class.DB(self.name, self.user, self.japd,
+                               self.host, self.port, auth_mech=self.auth_mech)
+
+        self.assertEqual(mongo.config, self.config)
 
     def test_db_attr(self):
 
@@ -190,7 +226,7 @@ class UnitTest(unittest.TestCase):
         mongo = mongo_class.DB(self.name, self.user, self.japd,
                                self.host, self.port)
 
-        self.assertEqual(mongo.config, self.config)
+        self.assertEqual(mongo.config, self.config2)
 
     def test_using_no_auth_db(self):